Removed the "Pseudo classes" and object parameterization from states/objects. This will reduce the memory overhead of BURLAP which previously was a cost that wasn't used by any of the BURLAP algorithms. Removing pseudo classes also removes a lot of ambiguity about whether an algorithm should reference the true class or the parameterized pseudo class.
If an algorithm wants to be able to use parameterized states (e.g., for
creating a parameterizable policy) it should handle parameterization
explicitly, or make use of a set of state features or hashing factory
that can be parameterized, thereby making the operation more clear to
developers.

Note: Action parameterization is different than state/object
parameterization. Action parameterization is a more clear feature and is
used by existing algorithms/domains and therefore remains in BURLAP.